requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.1
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.2
